requirements.txt
aiogram==3.15.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
APScheduler==3.10.4
